Fix Heatmap codegen, clean up code structure
JIRA: RAIL-4069
