+++ libs/sdk-ui-ext/src/internal/components/pluggableVisualizations/funnelChart/FunnelChartDescriptor.ts
@@ -9,8 +9,6 @@ import {
 import { PluggableFunnelChart } from "./PluggableFunnelChart";
 import { BaseChartDescriptor } from "../baseChart/BaseChartDescriptor";
 import {
-    chartAdditionalFactories,
-    chartConfigInsightConversion,
     filtersInsightConversion,
     getInsightToPropsConverter,
     getReactEmbeddingCodeGenerator,
@@ -18,6 +16,7 @@ import {
     singleAttributeBucketConversion,
     sortsInsightConversion,
 } from "../../../utils/embeddingCodeGenerator";
+import { chartAdditionalFactories, chartConfigInsightConversion } from "../chartCodeGenUtils";
 
 export class FunnelChartDescriptor extends BaseChartDescriptor implements IVisualizationDescriptor {
     public getFactory(): PluggableVisualizationFactory {
